% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\encoding{UTF-8}
\name{m_AddMortalidad}
\alias{m_AddMortalidad}
\title{Agregar Factor de Mortalidad a Tabla de Mortalidad}
\usage{
m_AddMortalidad(tbl, tasa = 0.1, raiz, edad_baja, edad_alta,
  genero_split = c(50, 50), metodo = "Lineal", redondeo = "Abajo")
}
\arguments{
\item{tbl}{Tabla de Mortalidad, de acuerdo con el método en m_CrearTblMortalidad}

\item{tasa}{Tasa de muertes por cada 100 mil}

\item{raiz}{Raíz de análisis (número de personas iniciales)}

\item{edad_baja}{Edad (excluyente) mínima para afectar con mortalidad}

\item{edad_alta}{Edad (excluyente) máxima para afectar con mortalidad}

\item{genero_split}{Objeto de dos números, en los que se divide la mortalidad (Hombre, Mujer)}

\item{metodo}{Metodo de división de mortalidad (Lineal, Normal, o pesos por cuartil).}

\item{redondeo}{Redondeo de mortalidad para Arriba o Abajo (default)}
}
\description{
Toma una tasa y distribuye de acuerdo a metodo a diferentes edades y generos
}
\details{
Para dividir por métodos de cuártil, en el parámetro se específica un objeto númerico que sume 100 entre todos. Tal
como c(10,40,50). En este caso, el total de las edades se dividirá en 3 (el largo del objeto) tomando el peso de 10%, 40% y 50%, cada tercio de rangos
de edad de acuerdo a mayor a menor. Si edad_baja = 10 y edad_alta = 40, de 10 a 20 se agregará el 10% de la mortalidad, de 20 a 30 el 40% y de 30 a 40 el 50%.
}
\examples{
tabla_nueva <- m_AddMortalidad(tabla_origen, 0.01, 100000, 12, 65, c(40,60), c(10,10,80), "Abajo")
}
\author{
Eduardo Flores
}
\seealso{
m_Metodos
}
\keyword{metodo}

