% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\encoding{UTF-8}
\name{m_CrearTblMortalidad}
\alias{m_CrearTblMortalidad}
\title{Crear Tabla de Mortalidad}
\usage{
m_CrearTblMortalidad(edad, defs, raiz)
}
\arguments{
\item{edad}{Vector de edades}

\item{defs}{Vector de defunciones, en mismo orden que edad}

\item{raiz}{Raíz de análisis (número de personas iniciales)}
}
\description{
A partir de edad, defunciones y raíz, crea una tabla de mortalidad hipotética.
}
\examples{
tabla <- m_CrearTblMortalidad(c(10,20,50), c(10,35, 90), 100)
}
\author{
Eduardo Flores
}
\seealso{
m_Metodos
}
\keyword{metodo}

